Подключил js как модули.
