+++ js/modules/util.js
@@ -1,0 +1,9 @@
+const getNumber = (minValue, maxValue, dotValue = 0) => {
+  if (maxValue >= minValue && minValue >= 0 && maxValue > 0) {
+    const randomFloat = Math.random() * (maxValue - minValue) + minValue;
+    return randomFloat.toFixed(dotValue);
+  }
+  return 'error message';
+};
+
+export { getNumber };
